move element selection useEffect to CadView
